Create Alternate Leaf Nodes Sum.cpp
You need to build a BST from the given array A of size N. The array contains distinct elements. You need to build the tree using the Array elements in the order of their arrival (A[0] becomes root and so on). Now, you need to sum the alternate leaf nodes of this BST and print the result.
NODE: Start from the "left side" of the tree when searching for leaf nodes. So sum leaftnode1+leaftnode3+leafnode5 ...